% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_statistics_module.R
\name{anova_server}
\alias{anova_server}
\title{Server for the univariate statistics module}
\usage{
anova_server(id)
}
\arguments{
\item{id}{module id}
}
\description{
Server for the univariate statistics module
}
