% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_base_FCs.R
\name{calc_base_FCs}
\alias{calc_base_FCs}
\title{caluclate base fold changes}
\usage{
calc_base_FCs(base_metabo, meta, data_stats)
}
\arguments{
\item{base_metabo}{untransformed metabolite matrix}

\item{meta}{sample data}

\item{data_stats}{statistics data}
}
\description{
caluclate base fold changes
}
