% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_and_normalization_module.R
\name{data_viewer_output}
\alias{data_viewer_output}
\title{metabolite data for data viewer tab}
\usage{
data_viewer_output(id)
}
\arguments{
\item{id}{module id}
}
\description{
metabolite data for data viewer tab
}
