% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_and_normalization_functions.R
\name{density_plot}
\alias{density_plot}
\title{density plot function}
\usage{
density_plot(metabo, meta, group)
}
\arguments{
\item{metabo}{metabolite matrix}

\item{meta}{sample data}

\item{group}{independent variable to group by}
}
\description{
density plot function
}
