% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_base_FCs.R
\name{get_base_data}
\alias{get_base_data}
\title{get base data function}
\usage{
get_base_data(base_metabo, meta, data_stats)
}
\arguments{
\item{base_metabo}{untransformed metabolite matrix}

\item{meta}{sample data}

\item{data_stats}{statistics data}
}
\description{
get base data function
}
