% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_img_functions.R
\name{get_path_image}
\alias{get_path_image}
\title{get pathway image function}
\usage{
get_path_image(path_list, path)
}
\arguments{
\item{path_list}{list of pathways}

\item{path}{selected pathway from path_list}
}
\description{
get pathway image function
}
