% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_and_normalization_functions.R
\name{histogram}
\alias{histogram}
\title{Histogram function}
\usage{
histogram(metabo, meta, metabolite, bins = 30, group = "none")
}
\arguments{
\item{metabo}{metabolite matrix}

\item{meta}{sample data}

\item{metabolite}{selected metabolites}

\item{bins}{number of bins to plot}

\item{group}{independent variable to group by}
}
\description{
Histogram function
}
