% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_and_normalization_module.R
\name{k_s_table_output}
\alias{k_s_table_output}
\title{Produces kurtosis and skewness table}
\usage{
k_s_table_output(id)
}
\arguments{
\item{id}{module id}
}
\description{
Produces kurtosis and skewness table
}
