% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG_gather_module.R
\name{KEGG_gather_server}
\alias{KEGG_gather_server}
\title{Server for the upload and transform data module}
\usage{
KEGG_gather_server(id)
}
\arguments{
\item{id}{module id}
}
\description{
Server for the upload and transform data module
}
