% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_and_normalization_functions.R
\name{kurtosis_and_skewness}
\alias{kurtosis_and_skewness}
\title{kurtosis and skewness function}
\usage{
kurtosis_and_skewness(metabo, meta, metabolite, group)
}
\arguments{
\item{metabo}{metabolite matrix}

\item{meta}{sample data}

\item{metabolite}{user selected metabolites}

\item{group}{independent variable to group by}
}
\description{
kurtosis and skewness function
}
