% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_and_normalization_functions.R
\name{metabolite_prevalence}
\alias{metabolite_prevalence}
\title{function to determine metabolite prevalence across samples}
\usage{
metabolite_prevalence(metabo, threshold)
}
\arguments{
\item{metabo}{metabolite matrix}

\item{threshold}{user supplied threshold for subsetting metabolites}
}
\description{
function to determine metabolite prevalence across samples
}
