% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG_gather_module.R
\name{pathway_image_output}
\alias{pathway_image_output}
\title{Function for synthetic pathway image}
\usage{
pathway_image_output(id)
}
\arguments{
\item{id}{module n}
}
\description{
Function for synthetic pathway image
}
