% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_and_normalization_module.R
\name{sample_viewer_output}
\alias{sample_viewer_output}
\title{sample data for data viewer tab}
\usage{
sample_viewer_output(id)
}
\arguments{
\item{id}{module id}
}
\description{
sample data for data viewer tab
}
