% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_statistics_module.R
\name{stats_viewer_output}
\alias{stats_viewer_output}
\title{Function for viewing stats data}
\usage{
stats_viewer_output(id)
}
\arguments{
\item{id}{module id}
}
\description{
Function for viewing stats data
}
