% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_and_normalization_module.R
\name{upload_and_normalization_server}
\alias{upload_and_normalization_server}
\title{Server for the upload and transform data module}
\usage{
upload_and_normalization_server(id)
}
\arguments{
\item{id}{module id}
}
\description{
Server for the upload and transform data module
}
