% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_and_normalization_module.R
\name{upload_and_normalization_UI}
\alias{upload_and_normalization_UI}
\title{UI for uploading and transforming data}
\usage{
upload_and_normalization_UI(id)
}
\arguments{
\item{id}{module id}
}
\description{
UI for uploading and transforming data
}
