% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_module.R
\name{varimpOutput}
\alias{varimpOutput}
\title{Function for variable importance output}
\usage{
varimpOutput(id)
}
\arguments{
\item{id}{module n}
}
\description{
Function for variable importance output
}
